Replace the term "Category" by "Type"
The Term "Categories" is mainly used in a thematical way. But on our platform, the "categories" define the type of a entry. Therefore I suggest to change this!
Theses Types may be extended in the future:
  - Organisations:
    - Initiatives (non-profit, movemets)
    - Companies (for-profit, services)
  - Events
Future Additional Categories could be
  - People
  - Products/Things
So we are not planning of having more than 5 Types